requirements.txt
fastapi[standard]>=0.110.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9

# Base de Datos (PostgreSQL)
sqlalchemy>=2.0.36
alembic>=1.13.0
psycopg[binary]>=3.2.0

# Seguridad y Autenticación
bcrypt>=4.0.0
PyJWT>=2.8.0

# Análisis estadístico (IDF - Gumbel)
scipy>=1.13.0
numpy>=1.26.0
